moving mock data into file, adjusting response handling
